Safari-611.1.10/LayoutTests/fullscreen/full-screen-api.html

<body>
    <script src="full-screen-test.js"></script>
<span></span>
<script>
    span = document.getElementsByTagName('span')[0];
    testExpected("document.webkitIsFullScreen", false);
    testExpected("document.webkitCancelFullScreen", undefined, "!=");
    testExpected("document.webkitCurrentFullScreenElement", null);
    testExpected("document.onwebkitfullscreenchange", null)
    testExpected("span.webkitRequestFullScreen", undefined, "!=");
    testExpected("span.onwebkitfullscreenchange", null) 
    endTest();
</script>
