third_party/blink/web_tests/platform/generic/fullscreen/full-screen-api-expected.txt

EXPECTED (document.webkitIsFullScreen == 'false') OK
EXPECTED (document.webkitCancelFullScreen != 'undefined') OK
EXPECTED (document.webkitCurrentFullScreenElement == 'null') OK
EXPECTED (document.onwebkitfullscreenchange == 'null') OK
EXPECTED (span.webkitRequestFullScreen != 'undefined') OK
EXPECTED (span.onwebkitfullscreenchange == 'null') OK
END OF TEST
